requirements.txt
fastapi
pydantic
python-telegram-bot
requests
typing